[feature] db transaction method
