+++ app/_layout.tsx
@@ -4,7 +4,8 @@ import { Text, View } from "react-native";
 import { useEffect, useState } from "react";
 import Colors from "@/constant/Color";
 import { CurrentActivityDetailContextProvider } from "@/context/CurrentActivityDetailContext";
-import { DB } from "@/utils/DB";
+import { DB } from "@/utils/db";
+import { init } from "@/utils/init";
 
 import {
   configureReanimatedLogger,
@@ -30,70 +31,7 @@ export default function RootLayout() {
       // if (__DEV__) {
       //   await DB.seed();
       // }
-      DB.register("expenses", "insert", (payload: any) => {
-        const aid = Array.isArray(payload)
-          ? payload[0].activity_id
-          : payload.activity_id;
-        DB.query(
-          `
-          UPDATE participants SET 
-          total_paid = (
-              SELECT COALESCE(SUM(e.amount), 0)
-              FROM expenses e
-              WHERE e.paid_by = participants.id
-          ) WHERE activity_id = ?
-        `,
-          [aid]
-        );
-      });
-      DB.register("participants", "insert", (payload: any) => {
-        const aid = Array.isArray(payload)
-          ? payload[0].activity_id
-          : payload.activity_id;
-        DB.query(
-          `
-          UPDATE participants SET
-          total_owed = (
-              SELECT COALESCE(SUM(
-                e.amount / (
-                    SELECT COUNT(*) 
-                    FROM participant_expenses pe2 
-                    WHERE pe2.expense_id = e.id
-                )
-              ), 0)
-              FROM expenses e
-              JOIN participant_expenses pe ON pe.expense_id = e.id
-              WHERE pe.participant_id = participants.id
-            )
-          WHERE activity_id = ?;
-        `,
-          [aid]
-        );
-      });
-      DB.register("participant_expenses", "insert", (payload: any) => {
-        const pid = Array.isArray(payload)
-          ? payload.map((pe) => pe.participant_id)
-          : [payload.participant_id];
-        DB.query(
-          `
-          UPDATE participants SET
-          total_owed = (
-              SELECT COALESCE(SUM(
-                e.amount / (
-                    SELECT COUNT(*) 
-                    FROM participant_expenses pe2 
-                    WHERE pe2.expense_id = e.id
-                )
-              ), 0)
-              FROM expenses e
-              JOIN participant_expenses pe ON pe.expense_id = e.id
-              WHERE pe.participant_id = participants.id
-            )
-          WHERE id IN (${pid.map(() => "?").join(", ")});
-        `,
-          pid
-        );
-      });
+      await init();
       setReady(true);
     } catch (err) {
       console.log(err);
